feat(db): add platform and timezone to user and daily user models
Adds `platform` and `timezone` fields to the `User` model to track device and location.
The `DailyUser` model is updated to include these fields and its `date` column
type is changed from `VARCHAR` to `DATE` with a `CURRENT_DATE` default.
The `add_to_daily_users` trigger is also updated to leverage these new fields.