refactor(user-tracking): streamline daily user tracking
Removed the `addToDailyUsers` method from dashboardService and the `add-to-daily-users` application event.
The logic for recording daily user activity is now integrated into the `setUserOnlineStatus` method in chatService, which updates the user's platform and timezone directly on the user model  which cause the add_daily_users trigger to be executed.
Additionally, date range queries for user growth trend and device/timezone statistics in dashboardService have been improved for accuracy.